Move reset_edit_handler_cache to wagtail.admin.panels.signal_handlers
